Depends on https://github.com/CitrineInformatics/platform-backend/pull/3003
The scope field in a custom scope:id pair can't contain
the delimiter string "::"

This is not a behavioral change, just a clarification in the docs.
Users can now include "::" in a custom ID field, which is a change
to the PR above, but the existing docs are framed around constraints.